+++ benchmark/WeihanLi.Common.Benchmark/CreateInstanceTest.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq.Expressions;
 using BenchmarkDotNet.Attributes;
+using WeihanLi.Common.Helpers;
 
 namespace WeihanLi.Common.Benchmark
 {
@@ -16,6 +17,12 @@ namespace WeihanLi.Common.Benchmark
         public MapperTest.B NewInstanceByReflection()
         {
             return Activator.CreateInstance<MapperTest.B>();
+        }
+
+        [Benchmark]
+        public MapperTest.B NewInstanceByActivatorHelper()
+        {
+            return ActivatorHelper.CreateInstance<MapperTest.B>(DependencyResolver.Current);
         }
 
         [Benchmark]